requirements.txt
patool
pytidylib